+++ jobbergate-cli/CHANGELOG.md
@@ -4,8 +4,10 @@ This file keeps track of all notable changes to jobbergate-cli
 
 ## Unreleased
 
+- Allow methods from `JobbergateBaseApplication` to return None for backward compatibility [ASP-4557]
 
 ## 4.2.1 -- 2024-01-15
+
 ## 4.2.0 -- 2024-01-08
 
 - Added instructions to checkbox questions [ASP-4042]